api/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
Pillow==9.5.0
python-dotenv==0.21.1
requests==2.28.2
urllib3==1.26.14
